delete comment dialog added

+++ src/app/components/posts/recipe/comments/comments.component.ts
@@ -1,4 +1,6 @@
 import { Component, Input } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from 'src/app/components/confirm-dialog/confirm-dialog.component';
 import { Comments } from 'src/app/models/models';
 import { PostService } from 'src/app/services/post.service';
 
@@ -13,15 +15,19 @@ export class CommentsComponent {
 
 
   constructor(
-    private postService: PostService
+    private postService: PostService,
+    private dialog: MatDialog
   ) { }
 
 
   deleteCom(comment: Comments): void {
-    this.postService.delete(comment.id as number).subscribe(() => {
-      // refresh comments
-      this.refreshComments();
-    });
+    this.dialog.open(ConfirmDialogComponent).afterClosed().subscribe((result: boolean) => {
+      if (!result) { return; }
+      this.postService.delete(comment.id as number).subscribe(() => {
+        // refresh comments
+        this.refreshComments();
+      });
+    })
   }
 
 }
